+++ branches/Cobalt-Task4708/MAC/APL/PAC/ICAL_Protocol/src/SubArray.cc
@@ -199,7 +199,7 @@ size_t SubArray::getSize() const
 
 size_t SubArray::pack(char* buffer) const
 {
-	size_t offset = 0;
+	uint32 offset = 0;
 
 	MSH_pack(buffer, offset, itsName);
 	MSH_pack(buffer, offset, itsAntennaSet);
@@ -211,7 +211,7 @@ size_t SubArray::pack(char* buffer) const
 
 size_t SubArray::unpack(const char* buffer)
 {
-	size_t offset = 0;
+	uint32 offset = 0;
 
 	MSH_unpack(buffer, offset, itsName);
 	MSH_unpack(buffer, offset, itsAntennaSet);
@@ -230,14 +230,14 @@ size_t SubArrayMap::getSize() const
 
 size_t SubArrayMap::pack(char* buffer) const
 {
-	size_t	offset = 0;
+	uint32	offset = 0;
 	MSH_pack(buffer, offset, (*this));
 	return (offset);
 }
 
 size_t SubArrayMap::unpack(const char* buffer)
 {
-	size_t offset = 0;
+	uint32 offset = 0;
 	MSH_unpack(buffer, offset, (*this));
 	return (offset);
 }
